Extract `BuildApiFlags::Flags`
Bug: b/424280895
Test: bazel run //cuttlefish/package:cvd -- fetch --default_build=git_main/aosp_cf_x86_64_only_phone-trunk_staging-userdebug --target_directory=$HOME/dl4 --host_substitutions=all
Test: bazel test '//...' --test_summary=terse
